Replace poi with vue-cli 🎉
